feat: making it impossible for a user to delete himself

+++ src/server/services/users.ts
@@ -13,6 +13,7 @@ import { UserApi } from "@/api/apis/user";
 import { pricingTiersQuotas, roleIds } from "@/server/utils/constants";
 import { getRequesterFromDecodedAccessToken, organizationMembersAreIdentical } from "@/db/users/helpers";
 import { getLatestBillingByOrganizationId } from "@/db/billing/operations";
+import { sameUsers } from "@/db/utils/helpers";
 
 export async function getMyOrganizationDetails(decodedAccessToken: CustomAuth0JwtPayload) {
     const requester = await getRequesterFromDecodedAccessToken(decodedAccessToken);
@@ -109,8 +110,9 @@ export async function deleteMyOrganizationMember(
     const managementAPIToken = await userApi.getManagementAPIToken();
     const memberToRemove = await getSpecificMemberOfTheOrganization(requester.organization_id, memberToDelete.email);
 
-    await userApi.deleteAuth0User(managementAPIToken, memberToRemove.auth0_id);
+    if (sameUsers(requester, memberToRemove)) return;
 
+    await userApi.deleteAuth0User(managementAPIToken, memberToRemove.auth0_id);
     memberToRemove.is_deleted = true;
     await removeUserByEmailFromActive(memberToRemove, memberToRemove.email);
 }
